almost finished

+++ Modules/Shop/database/migrations/2025_07_31_174212_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
 		Schema::create('products', function (Blueprint $table) {
 			$table->id();
 			$table->string('name');
-			$table->unsignedInteger('traffic_gb');
+			$table->decimal('traffic_gb', 10, 3);
 			$table->float('duration_days'); // به خاطر تست
 			$table->unsignedBigInteger('price');
 			$table->unsignedTinyInteger('user_limit')->default(1);
